requirements.txt
numpy
pandas
matplotlib
Pillow
moviepy
librosa
soundfile
tqdm
requests
python-dotenv

# modèles open-source
transformers>=4.30.0
torch==2.9.0 
torchaudio==2.9.0
openai-whisper   
sentence-transformers
transformers
accelerate 
bitsandbytes
ftfy
accelerate
huggingface_hub[hf_xet]
python-dotenv
openai
streamlit